started factoring out common cli util funcs
